+++ index.html
@@ -229,5 +229,6 @@
     </div>
   </template>
 
+  <script src="main.js" type="module"></script>
 </body>
 </html>
